% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPQ_Scoring_Function.R
\name{BPQ_raw_scoring}
\alias{BPQ_raw_scoring}
\title{BPQ_raw_scoring}
\usage{
BPQ_raw_scoring(df, form = "SF")
}
\arguments{
\item{df}{Data frame to be scored.}

\item{form}{Form of questionnaire. Original Form = "OF", Combined Autonomic Reactivity = "CReact", Short form = "SF", Supradiaphragmatic Reactivity = "Supra", Very Short Form = "VSF"}
}
\description{
Takes a data frame with columns named "1", "2", "3"... "46" representing the corresponding BPQ items.
First step of BPQ_all_scoring.
Returns a tibble (a form of data frame) of the BPQ subscores for each row of subject data.
}
\keyword{bpq}
