% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPQ_Scoring_Function.R
\name{BPQ_single_scoring}
\alias{BPQ_single_scoring}
\title{BPQ_single_scoring}
\usage{
BPQ_single_scoring(val, table)
}
\arguments{
\item{val}{Subscore value to be scored}

\item{table}{Table for reference}
}
\description{
Takes a subscore in the form of an integer and a table for the subscore to be referenced in. Function
is used by BPQ_per_t_scoring , but is usable independently.
Returns a factor containing the given score's percentile rank and t-score, respectively, in the given
table.
}
\keyword{bpq}
