% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRTable}
\alias{CRTable}
\title{2020 Population Norms for BPQ Combined Reactivity (Sub- and Supra-diaphragmatic)}
\format{
A data frame with 81 rows and 3 variables:
\describe{
\item{raw}{raw score, numeric}
\item{percents}{percentile score, numeric}
\item{tScores}{t scores, numeric}
...
}
}
\usage{
CRTable
}
\description{
A dataset containing population norms for BPQ Combined Reactivity, calculated from a sample collected in 2020
}
\keyword{datasets}
