% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SupTable_2018}
\alias{SupTable_2018}
\title{2018 Population Norms for BPQ Supra-diaphragmatic Reactivity}
\format{
A data frame with 61 rows and 3 variables:
\describe{
\item{raw}{raw score, numeric}
\item{percents}{percentile score, numeric}
\item{tScores}{t scores, numeric}
...
}
}
\usage{
SupTable_2018
}
\description{
A dataset containing population norms for BPQ Supra-diaphragmatic Reactivity, calculated from a sample collected in 2018
}
\keyword{datasets}
