% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Sample BPQ data to test functions.}
\format{
A data frame with 5 rows and 47 variables:
\describe{
\item{ID}{Respondent Id, Character String}
\item{bpq_1}{BPQ Item Response, Likert scale converted to numeric}
...
}
}
\usage{
test_data
}
\description{
A dataset containing dummy responses for BPQ item data
}
\keyword{datasets}
